+++ CHANGELOG.md
@@ -3,6 +3,7 @@
     * Added `onPreStep` callback
 * Item, Buff : `onDraw` callbacks now use `draw_actor` internally.
     * Added `onPreDraw` callbacks
+* Damager : Added `ATTACK_FLAG` and `set_attack_flags`
 
 ### v1.1.28
 * Instance : `exists` no longer throws an error if the argument is a string.